+++ src/pages/order/Order.jsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from 'react';
-import CommonLayout from '../components/layouts/CommonLayout';
-import { ColumnHeader } from '../components/common/customHeader';
-import MenuTab from '../components/order/MenuTab';
-import MenuList from '../components/order/MenuList';
-import OrderNavigationBar from '../components/order/OrderNavigationBar';
-import { OrderQueryService } from '../services/OrderService';
+import CommonLayout from '../../components/layouts/CommonLayout';
+import { ColumnHeader } from '../../components/common/customHeader';
+import MenuTab from '../../components/order/MenuTab';
+import MenuList from '../../components/order/MenuList';
+import OrderNavigationBar from '../../components/order/OrderNavigationBar';
+import { OrderQueryService } from '../../services/OrderService';
 
 function Order() {
     const [activeTab, setActiveTab] = useState('음료');
